+++ CHANGELOG.rst
@@ -4,5 +4,5 @@ Version 0.1.1 - Development
 * First official alpha release.
 * Five different type of applications: HTTP server, RPC server, distributed task queue,
   asynchronous test suite and asynchronous shell.
-* 4 examples spanning several functionalities.
-* **22 regression tests**
+* 4 examples spanning all applications.
+* **27 regression tests**